Fixing support for SSE2 on x86/x64, adding support for armv6l
Previously, compilation worked for all architectures at the
expense of the use of SSE2 instructions. The offending line
(where $$QMAKE_HOST.arch is checked) always returned true. This
is because $$QMAKE_HOST.arch was not read properly before the
equals() check, making the check return false. It was then not-ed
and returned true. Thus support for ARM was achieved because of
two errors cancelling eachother out, but negatively impacting
non-ARM processor performance. This is previously addressed in
https://github.com/DMDcoin/Diamond/commit/52be4f0c8b88a346f137be51e1f6019042e25383
because it added back in the -msse2 flag, but it added it for
all compiles instead of using it selectively for non-ARM
compiles.

This patch explicitly defines QMAKE_XCPUARCH = $$QMAKE_HOST.arch
before the equals check, removes the not on the equals() check,
and adds in a check for armv6l architecture CPUs as well.

It also adds support for armv6l (original Raspberry Pi) headless
wallet builds by adding in the appropriate check for xCPUARCH
in the Unix makefile.

All previously ARM-ified readmes have been merged into the main
readmes for cleanliness.